introduce AbstractLocalStateSecurity
