+++ x-pack/plugin/security/src/test/java/org/elasticsearch/xpack/security/LocalStateSecurity.java
@@ -16,13 +16,11 @@ import org.elasticsearch.injection.guice.Inject;
 import org.elasticsearch.license.LicenseService;
 import org.elasticsearch.license.XPackLicenseState;
 import org.elasticsearch.plugins.Plugin;
-import org.elasticsearch.plugins.ReloadablePlugin;
 import org.elasticsearch.protocol.xpack.XPackInfoRequest;
 import org.elasticsearch.protocol.xpack.XPackInfoResponse;
 import org.elasticsearch.protocol.xpack.XPackUsageRequest;
 import org.elasticsearch.threadpool.ThreadPool;
 import org.elasticsearch.transport.TransportService;
-import org.elasticsearch.xpack.core.LocalStateCompositeXPackPlugin;
 import org.elasticsearch.xpack.core.action.TransportXPackInfoAction;
 import org.elasticsearch.xpack.core.action.TransportXPackUsageAction;
 import org.elasticsearch.xpack.core.action.XPackInfoFeatureAction;
@@ -30,7 +28,6 @@ import org.elasticsearch.xpack.core.action.XPackInfoFeatureResponse;
 import org.elasticsearch.xpack.core.action.XPackUsageFeatureAction;
 import org.elasticsearch.xpack.core.action.XPackUsageFeatureResponse;
 import org.elasticsearch.xpack.core.action.XPackUsageResponse;
-import org.elasticsearch.xpack.core.security.SecurityExtension;
 import org.elasticsearch.xpack.core.ssl.SSLService;
 import org.elasticsearch.xpack.ilm.IndexLifecycle;
 import org.elasticsearch.xpack.monitoring.Monitoring;
@@ -39,7 +36,7 @@ import java.nio.file.Path;
 import java.util.Collections;
 import java.util.List;
 
-public class LocalStateSecurity extends LocalStateCompositeXPackPlugin implements ReloadablePlugin {
+public class LocalStateSecurity extends AbstractLocalStateSecurity {
 
     public static class SecurityTransportXPackUsageAction extends TransportXPackUsageAction {
         @Inject
@@ -102,21 +99,6 @@ public class LocalStateSecurity extends LocalStateCompositeXPackPlugin implement
                 return thisVar.getLicenseState();
             }
         });
-        plugins.add(new Security(settings, thisVar.securityExtensions()) {
-            @Override
-            protected SSLService getSslService() {
-                return thisVar.getSslService();
-            }
-
-            @Override
-            protected XPackLicenseState getLicenseState() {
-                return thisVar.getLicenseState();
-            }
-        });
-    }
-
-    protected List<SecurityExtension> securityExtensions() {
-        return List.of();
     }
 
     @Override
@@ -131,16 +113,5 @@ public class LocalStateSecurity extends LocalStateCompositeXPackPlugin implement
 
     public List<Plugin> plugins() {
         return plugins;
-    }
-
-    @Override
-    public void reload(Settings settings) throws Exception {
-        plugins.stream().filter(p -> p instanceof ReloadablePlugin).forEach(p -> {
-            try {
-                ((ReloadablePlugin) p).reload(settings);
-            } catch (Exception e) {
-                throw new RuntimeException(e);
-            }
-        });
     }
 }
